Modificacíones realizas al generador de batallas por modificaciones realizas a la administración del torneo a petición del frontend
