+++ src/main/java/com/mozcalti/gamingapp/controller/BatallasController.java
@@ -1,7 +1,7 @@
 package com.mozcalti.gamingapp.controller;
 
 import com.mozcalti.gamingapp.model.batallas.BatallasDTO;
-import com.mozcalti.gamingapp.service.batallas.CalendarizarEtapasTorneoService;
+import com.mozcalti.gamingapp.service.batallas.BatallasService;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -11,31 +11,31 @@ import org.springframework.web.bind.annotation.*;
 @AllArgsConstructor(onConstructor = @__(@Autowired))
 public class BatallasController {
 
-    private CalendarizarEtapasTorneoService calendarizarEtapasTorneoService;
+    private BatallasService batallasService;
 
     @GetMapping("/genera/{idEtapa}")
     public BatallasDTO getBatallasByIdEtapa(@PathVariable Integer idEtapa) {
-        return calendarizarEtapasTorneoService.generaBatallas(idEtapa);
+        return batallasService.generaBatallas(idEtapa);
     }
 
     @PostMapping("/save")
     public void saveBatallas(@RequestBody BatallasDTO batallasDTO) {
-        calendarizarEtapasTorneoService.saveBatallas(batallasDTO);
+        batallasService.saveBatallas(batallasDTO);
     }
 
     @GetMapping("/consulta/{idEtapa}")
     public BatallasDTO getBatallas(@PathVariable Integer idEtapa) {
-        return calendarizarEtapasTorneoService.getBatallas(idEtapa);
+        return batallasService.getBatallas(idEtapa);
     }
 
     @PutMapping(value = "/update")
     public void updateBatallas(@RequestBody BatallasDTO batallasDTO) {
-        calendarizarEtapasTorneoService.updateBatallas(batallasDTO);
+        batallasService.updateBatallas(batallasDTO);
     }
 
     @DeleteMapping("/delete/{idEtapa}")
     public void deleteBatallas(@PathVariable Integer idEtapa) {
-        calendarizarEtapasTorneoService.deleteBatallas(idEtapa);
+        batallasService.deleteBatallas(idEtapa);
     }
 
 }
