+++ src/main/java/com/mozcalti/gamingapp/repository/EtapaBatallaRepository.java
@@ -3,5 +3,10 @@ package com.mozcalti.gamingapp.repository;
 import com.mozcalti.gamingapp.model.EtapaBatalla;
 import org.springframework.data.repository.CrudRepository;
 
+import java.util.List;
+
 public interface EtapaBatallaRepository extends CrudRepository<EtapaBatalla, Integer> {
+
+    List<EtapaBatalla> findAllByIdEtapa(Integer idEtapa);
+
 }
