+++ src/main/java/com/mozcalti/gamingapp/service/batallas/BatallasService.java
@@ -4,6 +4,7 @@ import com.mozcalti.gamingapp.commons.GenericServiceAPI;
 import com.mozcalti.gamingapp.exceptions.ValidacionException;
 import com.mozcalti.gamingapp.model.BatallaParticipantes;
 import com.mozcalti.gamingapp.model.Batallas;
+import com.mozcalti.gamingapp.model.batallas.BatallasDTO;
 import com.mozcalti.gamingapp.model.catalogos.EtapasDTO;
 import com.mozcalti.gamingapp.model.catalogos.InstitucionDTO;
 import com.mozcalti.gamingapp.model.catalogos.ParticipanteDTO;
@@ -20,5 +21,15 @@ public interface BatallasService extends GenericServiceAPI<Batallas, Integer> {
     List<InstitucionDTO> getInstituciones() throws ValidacionException;
 
     List<ParticipanteDTO> getParticipantesByIdInstitucion(Integer idInstitucion) throws ValidacionException;
+
+    BatallasDTO generaBatallas(Integer idEtapa);
+
+    void saveBatallas(BatallasDTO batallasDTO) throws ValidacionException;
+
+    BatallasDTO getBatallas(Integer idEtapa);
+
+    void updateBatallas(BatallasDTO batallasDTO) throws ValidacionException;
+
+    void deleteBatallas(Integer idEtapa) throws ValidacionException;
 
 }
